+++ src/commands/Settings/AllowInlineCommands.js
@@ -1,6 +1,8 @@
 'use strict';
 
 const Command = require('../../Command.js');
+const { getChannel } = require('../../CommonFunctions.js');
+
 
 class AllowInlineCommands extends Command {
   constructor(bot) {
@@ -16,7 +18,7 @@ class AllowInlineCommands extends Command {
   async run(message) {
     let enable = message.strippedContent.match(this.regex)[1];
     const channelParam = message.strippedContent.match(this.regex)[2] ? message.strippedContent.match(this.regex)[2].trim().replace(/<|>|#/ig, '') : undefined;
-    const channel = this.getChannel(channelParam, message);
+    const channel = getChannel(channelParam, message, message.guild.channels);
     if (!enable) {
       const embed = {
         title: 'Usage',
@@ -40,26 +42,6 @@ class AllowInlineCommands extends Command {
     await this.bot.settings.setChannelSetting(channel, 'allowInline', allowInline);
     this.messageManager.notifySettingsChange(message, true, true);
     return this.messageManager.statuses.SUCCESS;
-  }
-
-  /**
-   * Get the list of channels to enable commands in based on the parameters
-   * @param {string|Array<Channel>} channelsParam parameter for determining channels
-   * @param {Message} message Discord message to get information on channels
-   * @returns {Array<string>} channel ids to enable commands in
-   */
-  getChannel(channelsParam, message) {
-    let { channel } = message;
-    if (typeof channelsParam === 'string') {
-      // handle it for strings
-      if (channelsParam !== 'here') {
-        channel = this.bot.client.channels.get(channelsParam.trim());
-      } else if (channelsParam === 'here') {
-        // eslint-disable-next-line prefer-destructuring
-        channel = message.channel;
-      }
-    }
-    return channel;
   }
 }
