+++ src/commands/Settings/RespondToSettings.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const Command = require('../../Command.js');
+const { getChannel } = require('../../CommonFunctions');
 
 class RespondToSettings extends Command {
   constructor(bot) {
@@ -15,7 +16,7 @@ class RespondToSettings extends Command {
   async run(message) {
     let enable = message.strippedContent.match(this.regex)[1];
     const channelParam = message.strippedContent.match(this.regex)[2] ? message.strippedContent.match(this.regex)[2].trim().replace(/<|>|#/ig, '') : undefined;
-    const channel = this.getChannel(channelParam, message);
+    const channel = getChannel(channelParam, message);
     if (!enable) {
       const embed = {
         title: 'Usage',
@@ -39,26 +40,6 @@ class RespondToSettings extends Command {
     await this.bot.settings.setChannelSetting(channel, 'respond_to_settings', enableResponse);
     this.messageManager.notifySettingsChange(message, true, true);
     return this.messageManager.statuses.SUCCESS;
-  }
-
-  /**
-   * Get the list of channels to enable commands in based on the parameters
-   * @param {string|Array<Channel>} channelsParam parameter for determining channels
-   * @param {Message} message Discord message to get information on channels
-   * @returns {Array<string>} channel ids to enable commands in
-   */
-  getChannel(channelsParam, message) {
-    let { channel } = message;
-    if (typeof channelsParam === 'string') {
-      // handle it for strings
-      if (channelsParam !== 'here') {
-        channel = this.bot.client.channels.get(channelsParam.trim());
-      } else if (channelsParam === 'here') {
-        // eslint-disable-next-line prefer-destructuring
-        channel = message.channel;
-      }
-    }
-    return channel;
   }
 }
