+++ src/commands/Settings/Stop.js
@@ -2,6 +2,8 @@
 
 const Command = require('../../Command.js');
 
+const { getChannel } = require('../../CommonFunctions.js');
+
 /**
  * Untrack an event or item
  */
@@ -18,32 +20,12 @@ class Untrack extends Command {
   async run(message) {
     const roomId = new RegExp('(?:\\<\\#)?\\d{15,}(?:\\>)?|here', 'ig');
     const channelParam = message.strippedContent.match(roomId) ? message.strippedContent.match(roomId)[0].trim().replace(/<|>|#/ig, '') : undefined;
-    const channel = this.getChannel(channelParam, message);
+    const channel = getChannel(channelParam, message, message.guild.channels);
 
     this.bot.settings.removeTypeNotifications(channel.id);
     this.bot.settings.removeItemNotifications(channel.id);
     this.messageManager.notifySettingsChange(message, true, true);
     return this.messageManager.statuses.SUCCESS;
-  }
-
-  /**
-   * Get the list of channels to enable commands in based on the parameters
-   * @param {string|Array<Channel>} channelsParam parameter for determining channels
-   * @param {Message} message Discord message to get information on channels
-   * @returns {Array<string>} channel ids to enable commands in
-   */
-  getChannel(channelsParam, message) {
-    let { channel } = message;
-    if (typeof channelsParam === 'string') {
-      // handle it for strings
-      if (channelsParam !== 'here') {
-        channel = this.bot.client.channels.get(channelsParam.trim());
-      } else if (channelsParam === 'here') {
-        // eslint-disable-next-line prefer-destructuring
-        channel = message.channel;
-      }
-    }
-    return channel;
   }
 }
